[test] notify when process already terminated (e.g. tango server crashed)

+++ tests/conftest.py
@@ -51,12 +51,21 @@ BEACON = [sys.executable, "-m", "bliss.config.conductor.server"]
 BEACON_DB_PATH = os.path.join(BLISS, "tests", "test_configuration")
 
 
+def eprint(*args):
+    print(*args, file=sys.stderr, flush=True)
+
+
 def wait_terminate(process):
+    cmd = repr(" ".join(process.args))
+    if process.poll() is not None:
+        eprint(f"Process {cmd} already terminated with code {process.returncode}")
+        return
     process.terminate()
     try:
         with gevent.Timeout(10):
             process.wait()
     except gevent.Timeout:
+        eprint(f"Process {cmd} doesn't finish: try to kill it ...")
         process.kill()
         with gevent.Timeout(10):
             process.wait()
@@ -114,7 +123,7 @@ def clean_gevent():
         except ReferenceError:
             continue
         if not ob.ready():
-            print(f"Dangling greenlet (setup): {ob}")
+            eprint(f"Dangling greenlet (setup): {ob}")
             e = RuntimeError(f"Dangling greenlet cannot be killed: {ob}")
             with gevent.Timeout(10, e):
                 ob.kill()
@@ -133,7 +142,7 @@ def clean_gevent():
         except ReferenceError:
             continue
         if not ob.ready():
-            print(f"Dangling greenlet (teardown): {ob}")
+            eprint(f"Dangling greenlet (teardown): {ob}")
         greenlets.append(ob)
     all_ready = all(gr.ready() for gr in greenlets)
     with gevent.Timeout(10, RuntimeError("Dangling greenlets cannot be killed")):
@@ -765,7 +774,7 @@ def tcp_listener(data_parser=None):
     finally:
         messages.put(StopIteration)
         for msg in messages:
-            print(f"\nUnvalidated message: {msg}")
+            eprint(f"Unvalidated message: {msg}")
         with gevent.Timeout(10):
             sock.close()
             glistener.kill()
